Major update. Uses 3.10.5.

+++ KineticGWT/src/net/edzard/kinetic/CustomShape.java
@@ -33,7 +33,7 @@ public class CustomShape extends Shape {
 	 * @param duration The time it will take for the animation to complete, in seconds
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(CustomShape target, int duration) {
+	public final Transition transitionTo(CustomShape target, double duration) {
 		return transitionTo(target, duration, null, null);
 	}
 	
@@ -46,7 +46,7 @@ public class CustomShape extends Shape {
 	 * @return An object for controlling the transition.
 	 */
 	// Delegates to shape (no special properties to animate)
-	public final Transition transitionTo(CustomShape target, int duration, EasingFunction ease, Runnable callback) {
+	public final Transition transitionTo(CustomShape target, double duration, EasingFunction ease, Runnable callback) {
 		StringBuffer sb = new StringBuffer();
 		return transitionToShape(target, sb, duration, ease, callback);
 	}
