+++ KineticGWT/src/net/edzard/kinetic/Image.java
@@ -121,7 +121,7 @@ public class Image extends Shape {
 	 * @param duration The time it will take for the animation to complete, in seconds
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(Image target, int duration) {
+	public final Transition transitionTo(Image target, double duration) {
 		return transitionTo(target, duration, null, null);
 	}
 	
@@ -134,7 +134,7 @@ public class Image extends Shape {
 	 * @param callback A function that will be called at the end of the animation
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(Image target, int duration, EasingFunction ease, Runnable callback) {
+	public final Transition transitionTo(Image target, double duration, EasingFunction ease, Runnable callback) {
 		StringBuffer sb = new StringBuffer();
 		
 		if (this.getWidth() != target.getWidth()) sb.append("width:").append(target.getWidth()).append(",");
