+++ KineticGWT/src/net/edzard/kinetic/Polygon.java
@@ -70,7 +70,7 @@ public class Polygon extends Shape {
 	 * @param duration The time it will take for the animation to complete, in seconds
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(Polygon target, int duration) {
+	public final Transition transitionTo(Polygon target, double duration) {
 		return transitionTo(target, duration, null, null);
 	}
 	
@@ -83,7 +83,7 @@ public class Polygon extends Shape {
 	 * @return An object for controlling the transition.
 	 */
 	// Delegates to shape (no special properties to animate)
-	public final Transition transitionTo(Polygon target, int duration, EasingFunction ease, Runnable callback) {
+	public final Transition transitionTo(Polygon target, double duration, EasingFunction ease, Runnable callback) {
 		StringBuffer sb = new StringBuffer();
 		return transitionToShape(target, sb, duration, ease, callback);
 	}
