+++ KineticGWT/src/net/edzard/kinetic/Rectangle.java
@@ -73,7 +73,7 @@ public class Rectangle extends Shape {
 	 * @param duration The time it will take for the animation to complete, in seconds
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(Rectangle target, int duration) {
+	public final Transition transitionTo(Rectangle target, double duration) {
 		return transitionTo(target, duration, null, null);
 	}
 	
@@ -85,7 +85,7 @@ public class Rectangle extends Shape {
 	 * @param callback A function that will be called at the end of the animation
 	 * @return An object for controlling the transition.
 	 */
-	public final Transition transitionTo(Rectangle target, int duration, EasingFunction ease, Runnable callback) {
+	public final Transition transitionTo(Rectangle target, double duration, EasingFunction ease, Runnable callback) {
 		StringBuffer sb = new StringBuffer();
 		if (this.getWidth() != target.getWidth()) sb.append("width:").append(target.getWidth()).append(",");
 		if (this.getHeight() != target.getHeight()) sb.append("height:").append(target.getHeight()).append(",");
